Moved Launcher and ScriptExec into the dev.equo.ide package.
