+++ solstice/src/main/java/dev/equo/ide/Launcher.java
@@ -11,7 +11,7 @@
  * Contributors:
  *     EquoTech, Inc. - initial API and implementation
  *******************************************************************************/
-package dev.equo.solstice;
+package dev.equo.ide;
 
 import java.io.BufferedOutputStream;
 import java.io.BufferedWriter;
@@ -27,8 +27,10 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
+import java.util.function.Consumer;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
+import javax.annotation.Nullable;
 
 /**
  * Thanks to Thipor Kong for his workaround for Gradle's windows problems.
@@ -62,12 +64,18 @@ public class Launcher {
 		if (blocking) {
 			return launchAndInheritIO(null, command);
 		} else {
-			ScriptExec.script(ScriptExec.quoteAll(command)).execSeparate();
+			ScriptExec.script(ScriptExec.quoteAll(command)).execSeparate(null);
 			return 0;
 		}
 	}
 
 	public static int launchAndInheritIO(File cwd, List<String> args)
+			throws IOException, InterruptedException {
+		return launchAndInheritIO(cwd, args, null);
+	}
+
+	public static int launchAndInheritIO(
+			File cwd, List<String> args, @Nullable Consumer<Process> monitorProcess)
 			throws IOException, InterruptedException {
 		var builder = new ProcessBuilder(args);
 		if (cwd != null) {
@@ -76,6 +84,13 @@ public class Launcher {
 		var process = builder.start();
 		var outPumper = new StreamPumper(process.getInputStream(), System.out);
 		var errPumper = new StreamPumper(process.getErrorStream(), System.err);
+		if (monitorProcess != null) {
+			new Thread(
+							() -> {
+								monitorProcess.accept(process);
+							})
+					.start();
+		}
 		int exitCode = process.waitFor();
 		process.getOutputStream().flush();
 		outPumper.join();
@@ -103,7 +118,7 @@ public class Launcher {
 					out.flush();
 				}
 			} catch (IOException e) {
-				throw Unchecked.wrap(e);
+				throw new RuntimeException(e);
 			}
 		}
 	}
